+++ Assets/RicKit/RFramework/CHANGELOG.md
@@ -1,4 +1,6 @@
 # Changelog
+## [1.4.6] - 2025-11-26
+- Add BindablePropertyExtension for ICollection<T>, IList<T>, HashSet<T>
 ## [1.4.5] - 2025-09-29
 - Add ForceNotify and UnRegisterAll to BindableProperty
 ## [1.4.4] - 2025-08-28
@@ -8,7 +10,7 @@
 ## [1.4.2] - 2025-06-12
 - refactor ICommand interfaces for improved execution handling
 ## [1.3.4] - 2025-06-03
-- refactor: improve service initialization logic in ServiceLocator and AsyncSupport
+- refactor: improve service initialization l~~~~ogic in ServiceLocator and AsyncSupport
 ## [1.3.1] - 2025-05-27
 - refactor event handling in ServiceLocator to use Dictionary<Type, Delegate>
 ## [1.3.0] - 2025-05-27